Save feature is working
Encodes the state of Model.Input to a string (encoded,
compressed, then converted to Base64) and stores this in the
URL query as "state". This is also decoded when the page is loaded.
This allows for restoring the state of the app by pasting in the
URL. It also prevents users from losing their state by
accidentally hitting "Back" in the browser.